requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.8
requests==2.32.3
python-dotenv==1.0.1
django-filter==24.3
